delete exception

+++ src/main/java/com/bootcamp/springjpa/services/UserService.java
@@ -4,10 +4,13 @@ import java.util.List;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Service;
 
 import com.bootcamp.springjpa.Entities.User;
 import com.bootcamp.springjpa.repositories.UserRepository;
+import com.bootcamp.springjpa.services.excepetion.DataBaseException;
 import com.bootcamp.springjpa.services.excepetion.ResourceNotFoundException;
 
 @Service
@@ -30,7 +33,18 @@ public class UserService {
 	}
 	
 	public void delete(Long id) {
-		userRepository.deleteById(id);
+		try {
+			
+			if(!userRepository.existsById(id)) {
+				throw new ResourceNotFoundException(id);	
+			} else {
+				userRepository.deleteById(id);
+				userRepository.flush();	
+			}
+			
+		} catch (DataIntegrityViolationException e) {
+			throw new DataBaseException(e.getMessage());
+		}
 	}
 	
 	public User update(Long id, User obj) {
